+++ src/widgets/systemtrayhelper.cpp
@@ -1,14 +1,19 @@
 #include "systemtrayhelper.h"
 
+#include <QAction>
 #include <QApplication>
 #include <QIcon>
 #include <QMenu>
+#include <QPointer>
 #include <QSystemTrayIcon>
 
 #include "mainwindow2.h"
+#include "viewarea2.h"
 #include "widgetsfactory.h"
+#include <controllers/viewareacontroller.h>
 #include <core/configmgr2.h>
 #include <core/coreconfig.h>
+#include <core/sessionconfig.h>
 #include <utils/widgetutils.h>
 
 using namespace vnotex;
@@ -45,6 +50,24 @@ QSystemTrayIcon *SystemTrayHelper::setupSystemTray(MainWindow2 *p_win,
                                [p_win]() { p_win->showMainWindow(); });
 
     WidgetUtils::addActionShortcutText(act, coreConfig.getShortcut(CoreConfig::Global_WakeUp));
+  }
+
+  {
+    auto *act = menu->addAction(MainWindow2::tr("Quick Note"));
+    act->setEnabled(false);
+    MainWindow2::connect(p_win->getViewArea(), &ViewArea2::corePropagationReady, act,
+                         [act]() { act->setEnabled(true); });
+    MainWindow2::connect(act, &QAction::triggered, p_win, [p_win, p_configMgr, act]() {
+      QPointer<QAction> actionGuard(act);
+      act->setEnabled(false);
+      if (p_configMgr->getSessionConfig().getQuickNoteSchemes().isEmpty()) {
+        p_win->showMainWindow();
+      }
+      p_win->getViewArea()->getController()->requestQuickNote();
+      if (actionGuard) {
+        actionGuard->setEnabled(true);
+      }
+    });
   }
 
   menu->addSeparator();